refactor: move route plugin
